# Implemented Cache System