--- a/ml-services/text-prediction-service/.env.example
+++ b/ml-services/text-prediction-service/.env.example
@@ -7,3 +7,8 @@ ENDPOINT = "/predict"
 
 # Model name
 MODEL= bert-base-uncased
+
+# Redis configuration
+REDIS_HOST=localhost
+REDIS_PORT=6379
+REDIS_TTL=3600  # Cache TTL in seconds (1 hour)