… ml-services/CACHING.md
… ml-services/sticker-generation-service/src/__pycache__/__init__.cpython-311.pyc
… ml-services/sticker-generation-service/src/__pycache__/__init__.cpython-313.pyc
… ml-services/sticker-generation-service/src/__pycache__/cached_sticker_generator.cpython-311.pyc
… ml-services/sticker-generation-service/src/__pycache__/disk_cache.cpython-311.pyc
… ml-services/sticker-generation-service/src/__pycache__/memory_lru_cache.cpython-311.pyc
… ml-services/sticker-generation-service/src/__pycache__/sticker_generator.cpython-311.pyc
… ml-services/sticker-generation-service/src/__pycache__/sticker_generator.cpython-313.pyc
… ml-services/sticker-generation-service/src/test_cache/sticker_A_cat_wearing_a_hat.png
… ml-services/sticker-generation-service/src/test_cache/sticker_A_dog_playing_basketball.png
… ml-services/sticker-generation-service/src/test_cache/sticker_A_monkey_typing_on_a_computer.png
… ml-services/sticker-generation-service/src/test_cache/sticker_An_elephant_painting_a_picture.png
… ml-services/text-prediction-service/src/__pycache__/__init__.cpython-311.pyc
… ml-services/text-prediction-service/src/__pycache__/__init__.cpython-313.pyc
… ml-services/text-prediction-service/src/__pycache__/cache.cpython-311.pyc
… ml-services/text-prediction-service/src/__pycache__/cache.cpython-313.pyc
… ml-services/text-prediction-service/src/__pycache__/cached_predictor.cpython-311.pyc
… ml-services/text-prediction-service/src/__pycache__/memory_cache.cpython-311.pyc
… ml-services/text-prediction-service/src/__pycache__/predictor.cpython-311.pyc
… ml-services/text-prediction-service/src/cached_predictor.py
… ml-services/text-prediction-service/src/main.py
… ml-services/text-prediction-service/src/memory_cache.py
… ml-services/text-prediction-service/src/predictor.py
… ml-services/text-prediction-service/src/test_cach/test_cache.py
… ml-services/text-prediction-service/src/test_cach/test_fallback_cache.py