# Feature: Cache System & Performance Optimization

## Overview
This pull request implements a comprehensive caching system for both the text prediction and sticker generation services, dramatically improving performance and reducing computational load.

## Key Features
- Two-tier caching strategy for both services
- Redis-based primary cache for text predictions
- In-memory and disk-based caching for sticker generation
- Robust fallback mechanisms
- Performance optimization through intelligent caching

## Implemented Caching Systems

### Text Prediction Service
- Primary Redis cache with configurable TTL
- In-memory fallback cache
- Automatic cache switching when Redis is unavailable
- Performance improvements from ~0.7-0.9s to ~0.001s for repeated predictions
![Screenshot 2025-03-23 at 23 15 07](https://github.com/user-attachments/assets/7f1d9761-1518-45f6-ad3a-e8aa49be90d6)
![Screenshot 2025-03-23 at 23 38 27](https://github.com/user-attachments/assets/be2cdb2b-6c30-43dc-a2d8-d72c0bbc86b9)
![Screenshot 2025-03-23 at 23 51 03](https://github.com/user-attachments/assets/b27622eb-15b0-437f-8f61-b4cf8d0fcf14)

### Sticker Generation Service
- Memory LRU cache for frequently accessed stickers
- Disk-based cache for persistent storage
- Mock generator for testing cache mechanisms
- Performance improvements from ~1-3s to ~0.05ms for cached items
![Screenshot 2025-03-24 at 15 32 08](https://github.com/user-attachments/assets/c73b9ac7-4c8b-4177-a626-8f0bfafa63db)

## Performance Metrics
- Text Predictions: 99.9% faster on cache hits
- Sticker Generation: >99.99% faster on memory cache hits
- Resilient to service interruptions with fallback mechanisms

## Testing
- Comprehensive test scripts for both services
- Validates cache hit/miss functionality
- Verifies fallback and performance optimization strategies

## Documentation
- Detailed `CACHING.md` explaining implementation, configuration, and best practices
- Inline code comments

## How to Test
1. Ensure Redis is running
2. Install dependencies:
   ```bash
   pip install redis Pillow torch transformers
   ```
3. Run test scripts:
   ```bash
   # Text Prediction Service
   python3 text-prediction-service/test_fallback_cache.py
   
   # Sticker Generation Service
   python3 sticker-generation-service/src/test_cache/test_sticker_cache.py
   ```

## Related Issues
- Closes #11 [Feature Request] Cache System & Performance Optimization

## Checklist
- [x] Implement Redis caching for text predictions
- [x] Create in-memory fallback cache
- [x] Implement disk-based caching for stickers
- [x] Write comprehensive test scripts
- [x] Create detailed documentation

## Notes
This implementation uses mock generators for testing.